+++ sdk/core/corehttp/tests/testserver_tests/coretestserver/coretestserver/test_routes/streams.py
@@ -39,6 +39,82 @@ def stream_compressed_header_error():
     yield b"test"
 
 
+def stream_jsonl_basic():
+    data = [
+        b'{"msg": "this is a message"}\n',
+        b'{"msg": "this is another message"}\n',
+        b'{"msg": "this is a third message"}\n{"msg": "this is a fourth message"}\n',
+    ]
+    yield from data
+
+
+def stream_jsonl_multiple_kv():
+    data = [
+        b'{"msg": "this is a hello world message", "planet": {"earth": "hello earth", "mars": "hello mars"}}\n',
+        b'{"msg": "this is a hello world message", "planet": {"venus": "hello venus", "jupiter": "hello jupiter"}}\n',
+    ]
+    yield from data
+
+
+def stream_jsonl_no_final_line_separator():
+    data = b'{"msg": "this is a message"}'
+    yield data
+
+
+def stream_jsonl_broken_up_data():
+    data = [b'{"msg": "this is a first message"}\n{"msg": ', b'"this is a second message"}\n']
+    yield from data
+
+
+def stream_jsonl_broken_up_data_cr():
+    data = [b'{"msg": "this is a first message"}\r\n{"msg": ', b'"this is a second message"}\r\n']
+    yield from data
+
+
+def stream_jsonl_invalid_data():
+    data = [b'{"msg": "this is a third m']
+    yield from data
+
+
+def stream_jsonl_escaped_newline_data():
+    data = b'{"msg": "this is a...\\nmessage"}\n'
+    yield data
+
+
+def stream_jsonl_escaped_broken_newline_data():
+    data = [b'{"msg": "this is a first message"}\n{"msg": "\\n', b'this is a second message"}\n']
+    yield from data
+
+
+def stream_jsonl_broken_incomplete_char():
+    data = [
+        b'{"msg": "this is a first message"}\n{"msg": "\xf0\x9d',
+        b"\x9c\x8bthis is a second message\xf0\x9d",
+        b'\x9c\x8b"}',
+        b'\n{"msg": "this is a third message"}',
+    ]
+    yield from data
+
+
+def stream_jsonl_list():
+    data = [
+        b'["this", "is", "a", "first", "message"]\n',
+        b'["this", "is", "a", "second", "message"]\n',
+        b'["this", "is", "a", "third", "message"]\n',
+    ]
+    yield from data
+
+
+def stream_jsonl_string():
+    data = [
+        b'"this"\n',
+        b'"is"\n',
+        b'"a"\n',
+        b'"message"\n',
+    ]
+    yield from data
+
+
 @streams_api.route("/basic", methods=["GET"])
 def basic():
     return Response(streaming_body(), status=200)
@@ -104,3 +180,60 @@ def upload():
             break
         byte_content += chunk
     return Response(byte_content, status=200)
+
+
+@streams_api.route("/jsonl_basic", methods=["GET"])
+def jsonl_basic():
+    return Response(stream_jsonl_basic(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_multiple_kv", methods=["GET"])
+def jsonl_multiple_kv():
+    return Response(stream_jsonl_multiple_kv(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_no_final_line_separator", methods=["GET"])
+def jsonl_no_final_line_separator():
+    return Response(stream_jsonl_no_final_line_separator(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_broken_up_data", methods=["GET"])
+def jsonl_broken_up_data():
+    return Response(stream_jsonl_broken_up_data(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_broken_up_data_cr", methods=["GET"])
+def jsonl_broken_up_data_cr():
+    return Response(stream_jsonl_broken_up_data_cr(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_invalid_data", methods=["GET"])
+def jsonl_invalid_data():
+    return Response(stream_jsonl_invalid_data(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_escaped_newline_data", methods=["GET"])
+def jsonl_escaped_newline_data():
+    return Response(stream_jsonl_escaped_newline_data(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_escaped_broken_newline_data", methods=["GET"])
+def jsonl_escaped_broken_newline_data():
+    return Response(
+        stream_jsonl_escaped_broken_newline_data(), status=200, headers={"Content-Type": "application/jsonl"}
+    )
+
+
+@streams_api.route("/jsonl_broken_incomplete_char", methods=["GET"])
+def jsonl_broken_incomplete_char():
+    return Response(stream_jsonl_broken_incomplete_char(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_list", methods=["GET"])
+def json_list():
+    return Response(stream_jsonl_list(), status=200, headers={"Content-Type": "application/jsonl"})
+
+
+@streams_api.route("/jsonl_string", methods=["GET"])
+def json_string():
+    return Response(stream_jsonl_string(), status=200, headers={"Content-Type": "application/jsonl"})